// apps/features/trap/arc/src/features/SecuritySettings/lib/types.ts
export type SecuritySettingsType = {
    aladdinId: string | null;
    price: number | null;
    modelFamily: string | null;
    interestRateScenario: string | null;
    overnightRisk: string | null;
    analysisDate: string | null;
    oadOacMultiplier: string | null;
    anserCode: string | null;
};